Implemented cutting off their stations before they're completed, tweaked ConnectedTile.tabulateScore(...)
We now terminate opponents' stations if they're within one of completion, but we don't do much in the way of extending our own routes intelligently.
There's still a lot of AI work to be done; a few ideas are recorded on the TODO lines in move(...).
The ConnectedTile.tabulateScore(...) method now only requires 1 parameter, making it logical for the first time in its history.