Implemented PlayerData.routeInDanger(...), introduced some potentially-broken additional AI
I added the routeInDanger(...) method, which may be used to check whether any enemy, a specific enemy, or we would be able to complete any given route.
I also added a defense mode to the AI, although I suspect it is severely broken.  I plan to re-implement the entire AI, making heavy use of the routeInDanger(...) method and the recently-added ability to removeTiles.  The current AI is likely most useful for its TODOs; in fact, it may not be helpful for much else at all.